Utils file + parenthesis tokenizer
I created an "utils.h" file in which i added a class that tokenizes the values in a parenthesis given in commands. The values are separated by spaces ' ' and by commas ','.
The class takes a value like:
(1, "John", "Ion")
and creates a string* array with the values 1, "John", "Ion" and counts them.